requirements.txt
# python == 3.12.10
torch==2.9.1
torchvision==0.24.1
numpy==2.3.5
matplotlib==3.10.8
scipy==1.16.3
pyyaml==6.0.3
opencv-python==4.13.0.92
scikit-learn==1.8.0  # tree for matching neighbours
trimesh==4.11.4  # visualization
fast-simplification==0.1.13